Move TitleCleaner and StaTask to utils namespace
The `TitleCleaner` and `StaTask` classes were moved from the `FoliCon.Modules` namespace to the `FoliCon.Modules.utils` namespace. This refactoring effort is meant to clean up the Modules directory better organization and readability of code.
